changed version back from int to long

+++ GroupF/src/domain/model/groupmembership/GroupMembership.java
@@ -16,7 +16,7 @@ public class GroupMembership extends DomainObject<Long> implements
 	private MembershipStatus status;
 	private Calendar lastUpdated;
 
-	public GroupMembership(Long id, int version, IUser member, IGroup group,
+	public GroupMembership(Long id, long version, IUser member, IGroup group,
 			MembershipStatus status, Calendar lastUpdated) {
 		super(id, version);
 		this.member = member;
